+++ concordium-std/CHANGELOG.md
@@ -8,6 +8,7 @@
   reverts the change in concordium-std 0.4.1.
 - Change the `receive_name` parameter of `HasActions::send` to use `ReceiveName`
   instead of `str`.
+- Rename `send` to `send_raw` in `HasActions`.
 - Add `simple_send`, a wrapper for `HasActions::send`, which automatically
   serializes `parameter` (using `Serial`).
 - Allow init and receive methods to return custom error codes that will be displayed to the user